libhdrplusservice: Implement pipeline and blocks
Implement a pipeline that can create streams and pipeline blocks
for HDR+ use case.

A PipelineBlock performs a lower level task given inputs and outputs.
Implemented SourceCaptureBlock, DummyProcessingBlock, and
CaptureResultBlock for dummy processing pipeline to verify
buffer flows in the pipeline. (No real processing yet.)

A PipelineStream keeps a queue of buffers of the same stream
configuration. The pipeline creates a route for each stream. The
input stream's route is SourceCaptureBlock -> DummyProcessingBlock.
The output stream's route is DummyProcessingBlock ->
CaptureResultBlock.

Bug: 28637233
Change-Id: Ie682ac61aa7e4ebd0908cec322de64130e670009