libhdrplusserver: Support circular input route
Simplify the routing logic so that there is one route for each block
data (input or output request that contains a set of buffers) instead
of one route for each buffer. This avoids a lot of sanity checks
because it doesn't support forking.

Support circular route for input so that input stream buffers can go
back to be captured for next frame when they are done.

Bug: 28637233
Change-Id: Idc7599f03a88dcc8630f67b61e59469d19b52c89