Customized liblog for easel.
Most binaries running on easel depend on liblog.so. Easel's liblog is
different from Android's liblog implementation. This liblog
implementation prints log to stdout and sends the log
to Android AP. To avoid module naming conflict, this module
is defined as libeasellog. When copying to easel's ramdisk,
this library needs to be renamed as liblog.so.

Bug: 34901928
Test: Tested on device
Change-Id: I54efac9a0a9c7e1812a54331acfa4fc1ccfe8ba2