libhdrplusservice: Handle frame drop in AP
Handle the case when AP drops a frame and doesn't send the frame
metadata to Easel.

Add a method to handle timeouts for waiting for events in a
block. In source capture block, if it times out, remove all
pending output results that have not received frame metadata for
a timeout duration.

Also when aborting an input or output request, send it back to the
first block if it has a circular route. Otherwise, send the
buffer back to the streams.

Test: GCA
Bug: 35399985
Change-Id: I1a0a45207397eb0bbbc5d040778c47bcab69f855