Separate server side libeasel to be libeaselservice.
libeaselservice is now a standalone library without dependencies.
In build time, binaries are still linked to libeasel that contains
both server and client code. On server side, libeaselservice
is installed as libeasel to replace build time libeasel and
contains only server code. This will avoid liblog to be double
linked on server runtime.

Test: Camera app on device.
Change-Id: I42048502cf35db7f2b39f903ea09ebc3ae240766