Adds logd.easel to read logs from /dev/socket/logdw.
This CL provides a way to multi-processing logging
for easel. The logging channel is EASEL_SERVICE_LOG,
separated for EASEL_SERVICE_SYSCTRL. Original liblog.so
from Android could be used on easel, which writes to
/dev/socket/logdw. When the IPC message is received
by logd.easel, it will send it over to Android AP for
logging. LogClient is the class implemented on AP side
to print logs to Android logcat.

Bug: 37756450
Test: Open camera app and check logs.
Change-Id: I0482582e3eb66061281562e18f657a0c6fff18d7