ezlsh: open server tty in receive thread
This fixes a race condition that could happen if ezlsh was started on
the client before it was started on the server. The server would get
CMD_OPEN_SHELL and CMD_TTY_DATA back-to-back and the shell session
thread wouldn't have opened the tty yet. With this change, we open the
tty while handling CMD_OPEN_SHELL.

Bug: 62617367
Test: ezlsh and Camera CTS
Change-Id: I82078ffcc4a93e996fe921a7601d8db80fb8f5a5