Add support for thermal thresholds and thermal conditions
EaselThermalMonitor now defines 5 thermal condition levels. An array of
thresholds is added for each desired thermal zone. Now, when the thermal
monitor runs, it checks each thermal zone and compares it to the
thresholds to determine the current thermal condition. The highest
(hottest) thermal condition is stored as the system thermal condition.

EaselControlServer now passes the thermal condition into
EaselClockControl. For now, the thermal condition is only used for
Functional mode since we already run in the lowest power configuration
for Bypass and Capture modes. I have selected four operating points for
Functional mode (see b/38426194) that balance average power and
processing latency.

The thresholds are still a WIP with characterization.

Bug: 32591563
Bug: 37328907
Bug: 38426194
Test: Camera CTS, lots of HDR+ shots
Change-Id: Ic7f2c4969a7f230f2c88a8f0b14adef64239304c