Improves ezlsh to support power on / off and check power state
Adds ezlsh poweron and ezlsh poweroff commands and only runs ezlsh
when Easel is powered on.

Make logClient enabled by default and don't wait for it to be started,
so 1) logs will always be printed and 2) ezlsh will not be blocked
on waiting for Easel to be turning on.

The only drawback is logs on easel side maybe missing if the logs
are sent early. But in practice, it will not make things worse
since even though logClient getting ready early, but the logs may
arrive after the ezlsh operation finishes. In those cases, the logs
may be missing anyway. However, in interactive shell session, logs
will not be missing, since there is enough time to send logs before
user turn down the session.

Test: ezlsh
Change-Id: I484025db697fe2aa7a24b42216bfb9131d09750f