Add libarect dependency to test.
These headers cannot be exported by other header
libraries because it is a static library.

This will be fixed in b/63599562.

When the vndk is enabled (BOARD_VNDK_VERSION=current),
vendor libraries must only link against vendor variants
in the build system. This also means that they receive
restricted sets of headers and that automatic global
headers are restricted.

Test: headers found with BOARD_VNDK_VERSION=current
Bug: 37342627
Change-Id: I42b29af5c567263866d4398a822b5c0e72803728