Add send API to support sending a group of buffers.
In NN API, client may need to send a group of buffers to server.
Currently we only support send one buffer at a time, and the buffer
has to be matched with a message. With the new added API, user
could just send the buffer over without providing dummy message.

Also an extra field ID is introduced in HardwareBuffer, which could
be used to note the buffer sequence in the group. The received buffer
object will share the same id as the output buffer.

Test: easelcomm2_test
Change-Id: Ia47625a2e2f41fe4c133ced0c03bf6652765cbee