Update MallocAHardwareBufferEaselLoopback
We find a bug when sending a large (>2KB) ion buffer from easel to
AP. If AP side receiving buffer is malloc based, AP will hang and
reboot (b/64398844). Now we modify the test server to receive and
loopback malloc buffer instead of ion buffer to temporarily
workaround the bug.

Test: easelcomm2_test
Bug: 64398844
Change-Id: I3fa50999047068e8185883ceca4548eadf058399