Hide implementation details of log_time struct
In the future, the sizes of tv_sec and tv_nsec (or even the size of
log_time struct itself) can change due to the 32-bit overflow expected
to happen in the year 2138. In order to hide such implementation details
to the clients of liblog, the two macros LOG_TIME_SEC and LOG_TIME_NSEC
are introduced.

Bug: 37629934
Test: build with BOARD_VNDK_VERSION=current
Change-Id: Icc36201a95e4ff1b62a45f574ed2906f027b3072