EaselControlClient: ignore return value when suspending
When we are trying to suspend, we are really just hoping to get to a
powered down state. So, any failures with easelcomm messages are not
important since our desired outcome is powering off. The existing code
had a corner only case where Easel experienced a fatal error (and
shutdown automatically). The camera HAL proceeded to suspend Easel, but
we failed because we couldn't send a Deactivate command. However,
sending this command is not really relevant since the goal is to
powerdown Easel.

Bug: 65653720
Test: Camera CTS and GCA
Change-Id: I834c7375331efd8d2465e080a26c4a24b76411f8