DO NOT MERGE pbcamera: Don't hold locks when sending events
Don't hold locks when sending messages to AP. Without this, there
may be a deadlock when AP submits a capture request while Easel
sends out a capture result and shutter callback.

Increase the number of output buffers to 3: 1 in capture result
block, 1 in HDR+ processing block, and 1 for next capture.

Test: CTS
Bug: 68403357
Change-Id: I6a0c79bcdb9707979d104db9d56b5cedb7673a66