EaselControl: Send deactivate cmd if no pending activate
If user does not call activate/deactivate, Easel stays in the initial
power state which is not bypass mode.

This patch sends a deactivate command at first if there is no pending
activate command.  Sending deactivate command is to save power; checking
pending activate command is to save startup latency.

This new behavior only applies to a process who owns an instance of
EaselControlClient.
This does not affect "ezlsh poweron".

Bug: 68007782
Test: GCA, EaselCamera, Camera CTS
Change-Id: I39ae36c914a06924c97d79c18a620676136d92c7