Move buffer into a class and add option to allocate buffer.
Add the option to allocate malloc buffer in HardwareBuffer and
the malloc buffer will be owned and managed by HardwareBuffer.
So in most cases, user don't need to worry about freeing
HardwareBuffer after use. Since HardwareBuffer is getting
complicated now, so it is moved to a separate file as a class.

Test: easelcomm2_test
Change-Id: I37de5b1065fc36491b4bd740035a12696823d215