HdrPlusClientTests: Fix multiple requests
libgcam will not pick the same frame as the base frame twice by looking
at the frame timestamp. When the test submits multiple requests, it
is possible that libgcam will not be able to pick any base frame because
all have been picked once. The change works around the issue by changing
the timestamps of input buffers before each requests.

Test: HdrPlusClientTest.CaptureMultiYuv with 30 requests
Bug: 67717682
Change-Id: I5304a6b79d251e2fca21a22369b66b9f02e6ddcb