Supports multiple model preparations.
Introduce a model id (int64) in PaintboxDriver. Every model prepared
is assigned with a model id. The model id is monotonously increasing
and will not be reused. A model mapi mapping model id to the model
object is added to both client and server. To execute a request,
the model should be present in the map first.

Bug: 69673074
Test: NeuralNetworksOemTest
Change-Id: Id3f2eade40501109116f50d256b8d5096691eb4e